+++ include/avif/internal.h
@@ -34,8 +34,10 @@ uint32_t avifNTOHL(uint32_t l);
 uint64_t avifHTON64(uint64_t l);
 uint64_t avifNTOH64(uint64_t l);
 
-int avifFullToLimited(int depth, int v);
-int avifLimitedToFull(int depth, int v);
+int avifFullToLimitedY(int depth, int v);
+int avifFullToLimitedUV(int depth, int v);
+int avifLimitedToFullY(int depth, int v);
+int avifLimitedToFullUV(int depth, int v);
 
 void avifCalcYUVCoefficients(avifImage * image, float * outR, float * outG, float * outB);